resolve lifetime issue
